+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -441,6 +441,7 @@ sub build_isf_public_interface_contract {
             't/1272-isf-enum-member-rule-values.t',
             't/1273-isf-enum-member-rule-expression-values.t',
             't/1274-isf-enum-member-rule-guard-values.t',
+            't/1275-isf-enum-member-condition-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',